paddle_Bash_Arcada game example added
I added an example game called paddle_Bash_Arcada. I had already added this game to the ArduboyTones for Arcada Fork.

But I have now created a version that uses native Arduino tones rather than ArduboyTones, so this version is appropriate for the base Adafruit_Arcada library. It is otherwise the same game. I wrapped a sound class (within the example code) around the native Arduino tones to make it easy to play non-blocking sounds in game loops (without resorting to ArduboyTones). The Arcada callback timer is used to update the sound using the function from the sound class.

I have tested the code on a Pygamer and it is working.